1.6.4: Now operational, a few known bugs left
Converted from BC to TE3 power APIs
Entirely removed BC power API
Fixed UI paths so UI now renders properly
Fixed lots of crashes
Cleaned up power implementation internally, finalized conversion to
floats

+++ erogenousbeef/bigreactors/common/tileentity/TileEntityCyaniteReprocessor.java
@@ -70,13 +70,13 @@ public class TileEntityCyaniteReprocessor extends TileEntityPoweredInventoryFlui
 	}
 
 	@Override
-	public float getMaxEnergyStored() {
-		return 1000f;
+	public int getMaxEnergyStored() {
+		return 1000;
 	}
 
 	@Override
-	public float getCycleEnergyCost() {
-		return 100f;
+	public int getCycleEnergyCost() {
+		return 100;
 	}
 
 	@Override
